+++ src/util/hash.h
@@ -8,6 +8,6 @@
  * 
  * Adapted from: http://www.cse.yorku.ca/~oz/hash.html
 */
-uint64_t djb2_hash_string(unsigned char *str);
+uint32_t djb2_hash_string(unsigned char *str);
 
 #endif
